requirements.txt
appnope==0.1.4
asttokens==3.0.0
beautifulsoup4==4.13.5
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
distlib==0.4.0
executing==2.2.0
filelock==3.19.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
parso==0.8.4
pexpect==4.9.0
pillow==12.0.0
pipenv==2025.0.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.0.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.43
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
wcwidth==0.2.13
Werkzeug==3.1.3
